chore: move widget to components directory
