+++ lib/app/features/chat/views/pages/share_via_message_modal/components/share_options.dart
@@ -10,8 +10,8 @@ import 'package:ion/app/components/separated/separated_row.dart';
 import 'package:ion/app/extensions/asset_gen_image.dart';
 import 'package:ion/app/extensions/build_context.dart';
 import 'package:ion/app/extensions/num.dart';
-import 'package:ion/app/extensions/theme_data.dart';
 import 'package:ion/app/features/auth/providers/auth_provider.c.dart';
+import 'package:ion/app/features/chat/views/pages/share_via_message_modal/components/share_options_menut_item.dart';
 import 'package:ion/app/features/chat/views/pages/share_via_message_modal/components/share_post_to_story_content.dart';
 import 'package:ion/app/features/feed/data/models/entities/modifiable_post_data.c.dart';
 import 'package:ion/app/features/feed/providers/feed_bookmarks_notifier.c.dart';
@@ -55,7 +55,7 @@ class ShareOptions extends HookConsumerWidget {
           separator: SizedBox(width: 12.0.s),
           children: [
             if (isPost)
-              _ShareOptionsMenuItem(
+              ShareOptionsMenuItem(
                 buttonType: ButtonType.primary,
                 icon: isCapturing.value
                     ? const CenteredLoadingIndicator()
@@ -70,7 +70,7 @@ class ShareOptions extends HookConsumerWidget {
                           mediaService,
                         ),
               ),
-            _ShareOptionsMenuItem(
+            ShareOptionsMenuItem(
               buttonType: ButtonType.dropdown,
               icon: Assets.svg.iconBlockCopy1.icon(size: iconSize, color: Colors.black),
               label: context.i18n.feed_copy_link,
@@ -80,7 +80,7 @@ class ShareOptions extends HookConsumerWidget {
               },
             ),
             if (isPost)
-              _ShareOptionsMenuItem(
+              ShareOptionsMenuItem(
                 buttonType: ButtonType.dropdown,
                 icon: Assets.svg.iconBookmarks.icon(size: iconSize, color: Colors.black),
                 label: context.i18n.button_bookmark,
@@ -93,25 +93,25 @@ class ShareOptions extends HookConsumerWidget {
                   }
                 },
               ),
-            _ShareOptionsMenuItem(
+            ShareOptionsMenuItem(
               buttonType: ButtonType.dropdown,
               icon: Assets.svg.iconFeedWhatsapp.icon(size: iconSize),
               label: context.i18n.feed_whatsapp,
               onPressed: () {},
             ),
-            _ShareOptionsMenuItem(
+            ShareOptionsMenuItem(
               buttonType: ButtonType.dropdown,
               icon: Assets.svg.iconFeedTelegram.icon(size: iconSize),
               label: context.i18n.feed_telegram,
               onPressed: () {},
             ),
-            _ShareOptionsMenuItem(
+            ShareOptionsMenuItem(
               buttonType: ButtonType.dropdown,
               icon: Assets.svg.iconLoginXlogo.icon(size: iconSize),
               label: context.i18n.feed_x,
               onPressed: () {},
             ),
-            _ShareOptionsMenuItem(
+            ShareOptionsMenuItem(
               buttonType: ButtonType.dropdown,
               icon: Assets.svg.iconFeedMore.icon(size: iconSize),
               label: context.i18n.feed_more,
@@ -188,42 +188,5 @@ class ShareOptions extends HookConsumerWidget {
         isCapturing.value = false;
       }
     }
-  }
-}
-
-class _ShareOptionsMenuItem extends StatelessWidget {
-  const _ShareOptionsMenuItem({
-    required this.icon,
-    required this.label,
-    required this.onPressed,
-    required this.buttonType,
-  });
-
-  final Widget icon;
-  final String label;
-  final VoidCallback onPressed;
-  final ButtonType buttonType;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 70.0.s,
-      child: Column(
-        children: [
-          Button.icon(
-            type: buttonType,
-            onPressed: onPressed,
-            icon: icon,
-          ),
-          SizedBox(height: 6.0.s),
-          Text(
-            label,
-            style: context.theme.appTextThemes.caption2,
-            textAlign: TextAlign.center,
-            overflow: TextOverflow.ellipsis,
-          ),
-        ],
-      ),
-    );
   }
 }
