+++ lib/app/features/chat/views/pages/share_via_message_modal/components/share_options_menut_item.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+import 'package:ion/app/components/button/button.dart';
+import 'package:ion/app/extensions/extensions.dart';
+
+class ShareOptionsMenuItem extends StatelessWidget {
+  const ShareOptionsMenuItem({
+    required this.icon,
+    required this.label,
+    required this.onPressed,
+    required this.buttonType,
+    super.key,
+  });
+
+  final Widget icon;
+  final String label;
+  final VoidCallback onPressed;
+  final ButtonType buttonType;
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: 70.0.s,
+      child: Column(
+        children: [
+          Button.icon(
+            type: buttonType,
+            onPressed: onPressed,
+            icon: icon,
+          ),
+          SizedBox(height: 6.0.s),
+          Text(
+            label,
+            style: context.theme.appTextThemes.caption2,
+            textAlign: TextAlign.center,
+            overflow: TextOverflow.ellipsis,
+          ),
+        ],
+      ),
+    );
+  }
+}
